Update description of afw.
The fundamental change here is removing the display of the Exposure class from
the UML model, which is no longer either accurate or helpful.

Also removed "todo list" of improvements required to the prototype, since it
was obviously incomplete and also based on a snapshot at one point in time
rather than an overarching design.

My current thinking is that attempting to itemize everything we might try to
cram into afw over the course of construction is impractical, so we have to
leave it to be driven by the more concrete requirements in other WBSs, and
that's effectively what I've said here. Happy to be contradicted.