Begin rewrite of introduction.
This removes a lot of soon-to-be-obsolete, mostly-redundant text,
while adding a brief outline for a more concise overview of the rest
of the document.

Intent as that each Production section will be start with a brief
overview of that section (centered around a very high-level diagram),
so there's no need for the longer summaries here.