Restructure DRP pipelines to handle CalExp updates
This adds a new figure to the coaddition and diffim section, using it
to structure the text and describe the uncertain data flow surrounding
the many steps that both depend on and update CalExp components.

Ideally, we'd go back and restructure the ImChar/JointCal section more
with this in mind (there's a bit of an imbalance now in calling out the
distinction between flavors of ImChar and JointCal and not doing that
with potential non-final DiffIm flavors).