Add cross-references for coadd scenario bullets.
This required defining a macro that seems to limit what can go in
a description environment item (it didn't like one I had with escaped
curly braces), but I think that's enough of an edge case that it's a
worthwhile tradeoff.  Someone with better latex-fu than me may be able
to fix that as well.